feat: use react-tracked on settings
