+++ packages/dapp/src/AppProvider.tsx
@@ -1,10 +1,11 @@
 import { ArcxAnalyticsProvider } from '@arcxmoney/analytics';
 import { CssBaseline } from '@mui/material';
-import { defaultSettings, SettingsProvider } from '@transferto/shared/src';
 import type { PropsWithChildren } from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { BackgroundGradient } from './components/BackgroundGradient';
 import { HistoryMigration } from './components/HistoryMigration';
+import { SettingsProvider } from './providers';
+
 import { ChainInfosProvider } from './providers/ChainInfosProvider';
 import { I18NProvider } from './providers/I18nProvider';
 import { MenuProvider } from './providers/MenuProvider';
@@ -18,7 +19,7 @@ export const AppProvider: React.FC<PropsWithChildren<{}>> = ({ children }) => {
     <QueryClientProvider client={queryClient}>
       <ChainInfosProvider>
         <I18NProvider>
-          <SettingsProvider defaultSettings={defaultSettings}>
+          <SettingsProvider>
             <MenuProvider>
               <ThemeProvider>
                 <ArcxAnalyticsProvider
