+++ packages/dapp/src/const/MenuConfig/ConnectedMenu/Connected-MenuItems.tsx
@@ -6,12 +6,11 @@ import ContentCopyIcon from '@mui/icons-material/ContentCopy';
 import PowerSettingsNewIcon from '@mui/icons-material/PowerSettingsNew';
 import { Avatar } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
-import { useSettings } from '@transferto/shared/src/hooks';
 import { walletDigest } from '@transferto/shared/src/utils/walletDigest';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { SubMenuKeys } from '../../../const';
-import { useMenu } from '../../../hooks';
+import { useMenu, useSettings } from '../../../hooks';
 import { useChainInfos } from '../../../providers/ChainInfosProvider';
 import { useWallet } from '../../../providers/WalletProvider';
 import { MenuListItem, TWallets } from '../../../types';
@@ -19,7 +18,7 @@ import { MenuListItem, TWallets } from '../../../types';
 const ConnectedMenuItems = () => {
   const { t: translate } = useTranslation();
   const i18Path = 'navbar.walletMenu.';
-  const settings = useSettings();
+  const { onWalletDisconnect } = useSettings();
   const { menu, onCopyToClipboard, onCloseAllNavbarMenus } = useMenu();
 
   const theme = useTheme();
@@ -93,7 +92,7 @@ const ConnectedMenuItems = () => {
       onClick: () => {
         disconnect();
         onCloseAllNavbarMenus();
-        settings.onWalletDisconnect();
+        onWalletDisconnect();
       },
     },
   ];
