+++ packages/dapp/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useMenu';
+export * from './useSettings';
 export * from './useUserTracking/';
