+++ packages/dapp/src/providers/ThemeProvider.tsx
@@ -1,11 +1,11 @@
 import { useMediaQuery } from '@mui/material';
 import { ThemeProvider as MuiThemeProvider } from '@mui/material/styles';
-import { useSettings } from '@transferto/shared/src/hooks';
+import { useSettings } from '@transferto/dapp/src/hooks';
 import { darkTheme, lightTheme } from '@transferto/shared/src/theme';
 import React, { PropsWithChildren, useMemo } from 'react';
 
 export const useDetectDarkModePreference = () => {
-  const settings = useSettings();
+  const { settings } = useSettings();
   const isDarkModeHook = useMediaQuery('(prefers-color-scheme: dark)');
   if (!settings.themeMode || settings.themeMode === 'auto') {
     return !!isDarkModeHook ? true : false;
