+++ packages/dapp/src/providers/index.ts
@@ -1,0 +1,6 @@
+export * from './ChainInfosProvider';
+export * from './I18nProvider';
+export * from './MenuProvider';
+export * from './SettingsProvider';
+export * from './ThemeProvider';
+export * from './WalletProvider';
