--- packages/shared/src/contexts/index.ts
@@ -1,1 +1,0 @@
-export * from './SettingsContext';
