+++ packages/shared/src/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from './use-settings';
 export * from './use-translation';
