--- packages/shared/src/hooks/use-settings.ts
@@ -1,6 +1,0 @@
-import { useContext } from 'react';
-import { SettingsContext } from '../contexts/SettingsContext';
-
-// ----------------------------------------------------------------------
-
-export const useSettings = () => useContext(SettingsContext);
